Search users settings switching between sidebar/topbar nav on screen width change. Modals for topbar nav showing right data, but not functioning correctly yet
